refactor(Ship): change ShipId type from int to long
  Class ShipInfo: ~change ShipId type (support larger IDs)
  Class ShipTransaction: ~change ShipId type (support larger IDs)
  Class GraphShipItem: ~change ShipId type (support larger IDs)

refactor(Transaction): change ShipId type from int to long

  Class Transaction: ~change ShipId type (support larger IDs)

refactor(ChartPalette): add PickForLong method

  Class ChartPalette: +added PickForLong method (support long keys)

refactor(ShipsDataTransactionsModel): change ShipId type from int to long

  Class ShipsDataTransactionsModel: ~change ships dictionary key type (support larger IDs)

refactor(ShipsGraphTradesModel): change ShipId type from int to long

  Class ShipsGraphTradesModel: ~change LoadCumulativeProfitPoints method signature (support long ShipId)

refactor(ShipsGraphsBaseModel): change ShipId type from int to long

  Class ShipsGraphsBaseModel: ~change active ship IDs and series dictionaries to long (support larger IDs)

refactor(StatsShipsLoadBaseModel): change ShipId type from int to long

  Class StatsShipsLoadBaseModel: ~change LoadEntries method signature (support larger IDs)

refactor(StatsShipsLoadTradesModel): change ShipId type from int to long

  Class StatsShipsLoadTradesModel: ~change LoadEntries method signature (support larger IDs)

refactor(StatsShipsLoadTransactionsModel): change ShipId type from int to long

  Class StatsShipsLoadTransactionsModel: ~change LoadEntries method signature (support larger IDs)

refactor(StatsShipsWaresBaseModel): change ShipId type from int to long

  Class StatsShipsWaresBaseModel: ~change LoadData method signature (support larger IDs)

refactor(StatsShipsWaresTradesModel): change ShipId type from int to long

  Class StatsShipsWaresTradesModel: ~change LoadData method signature (support larger IDs)

refactor(StatsShipsWaresTransactionsModel): change ShipId type from int to long

  Class StatsShipsWaresTransactionsModel: ~change LoadData method signature (support larger IDs)

refactor(StatsWaresShipsBaseModel): change ShipId type from int to long

  Class StatsWaresShipsBaseModel: ~change LoadData method signature (support larger IDs)

refactor(StatsWaresShipsTradesModel): change ShipId type from int to long

  Class StatsWaresShipsTradesModel: ~change LoadData method signature (support larger IDs)

refactor(StatsWaresShipsTransactionsModel): change ShipId type from int to long

  Class StatsWaresShipsTransactionsModel: ~change LoadData method signature (support larger IDs)

+++ src/Models/Ship.cs
@@ -4,14 +4,14 @@ namespace X4PlayerShipTradeAnalyzer.Models;
 
 public class ShipInfo
 {
-  public int ShipId { get; set; }
+  public long ShipId { get; set; }
   public string? ShipName { get; set; }
   public decimal? EstimatedProfit { get; set; }
 }
 
 public class ShipTransaction
 {
-  public int ShipId { get; set; }
+  public long ShipId { get; set; }
 
   // raw numeric fields for aggregation
   public long RawTime { get; set; }
@@ -30,7 +30,7 @@ public class ShipTransaction
 
 public class GraphShipItem : INotifyPropertyChanged
 {
-  public int ShipId { get; init; }
+  public long ShipId { get; init; }
   public string ShipName { get; init; } = string.Empty;
 
   protected decimal _estimatedProfit;
